Updated List Protocol and Strict Commands
Updated the message sent to the client in the list protocol to ensure it includes the radius or sides values in message. Also, forgot to mention in previous commit, but the command check is updated to be case sensitive and more strict, this means the command must always be in uppercase or the server will consider it as an invalid command.